WEB1-87 WEB1-89 WEB1-126 WEB1-127
Tested the functionality of the NASA APOD and the phone number feature. You can now only type in 10 numbers into the phone number bar and the sight seeing page is changed every day. Added the TEST-89-3 results in a separate folder due to an error based of off my internet connection that makes it only work once in a while.